requirements.txt
# Core LLM and AI
openai>=1.3.0,<2.0.0
anthropic>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.5

# GraphQL and API clients
graphql-core>=3.2.0
gql>=3.4.0
aiohttp>=3.8.0
requests>=2.31.0

# Blockchain and Web3
web3>=6.0.0
moralis>=0.1.0
eth-account>=0.9.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
tabulate>=0.9.0
pydantic>=2.0.0
asyncio>=3.4.3
redis>=4.5.0
celery>=5.3.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
